refactor: update rules
