+++ package.json
@@ -16,6 +16,7 @@
     "codegen": "pnpm --filter=@cgr/codegen codegen"
   },
   "devDependencies": {
+    "@cgr/config": "workspace:*",
     "@changesets/changelog-github": "^0.4.8",
     "@changesets/cli": "^2.26.0",
     "@trivago/prettier-plugin-sort-imports": "^4.0.0",
