+++ packages/config/eslint-base.js
@@ -1,0 +1,35 @@
+// @ts-check
+
+/** @type {import('eslint').Linter.Config} */
+const config = {
+  root: true,
+  parser: '@typescript-eslint/parser',
+  plugins: ['@typescript-eslint/eslint-plugin'],
+  extends: [
+    'eslint:recommended',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:prettier/recommended',
+  ],
+  rules: {
+    '@typescript-eslint/no-empty-function': 'off',
+  },
+  overrides: [
+    {
+      files: ['*.js', '*.cjs'],
+      rules: {
+        '@typescript-eslint/no-var-requires': 'off',
+        'no-undef': 'off',
+      },
+    },
+  ],
+  settings: {
+    react: {
+      version: '18.2.0',
+    },
+    next: {
+      rootDir: ['apps/*/'],
+    },
+  },
+}
+
+module.exports = config
