+++ packages/config/eslint-next.js
@@ -1,0 +1,11 @@
+// @ts-check
+
+const base = require('./eslint-base')
+
+/** @type {import('eslint').Linter.Config} */
+const config = {
+  ...base,
+  extends: [...(base.extends ?? []), 'next/core-web-vitals'],
+}
+
+module.exports = config
