+++ packages/config/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@cgr/config",
+  "version": "0.0.0",
+  "private": true
+}
